+++ app/src/app.ts
@@ -1,4 +1,5 @@
-import { NegociacaoController } from "./controllers/negociacao-controller.js";
+import { NegociacaoController } from './controllers/negociacao-controller.js';
+
 const controller = new NegociacaoController();
 const form = document.querySelector('.form');
 if (form) {
@@ -6,7 +7,6 @@ if (form) {
         event.preventDefault();
         controller.adiciona();
     });
-}
-else {
+} else {
     throw Error('Não foi possível inicializar a aplicação. Verifique se o form existe.');
 }
