+++ app/src/controllers/negociacao-controller.ts
@@ -1,8 +1,8 @@
-import { DiasDaSemana } from "../enums/dias-da-semana.js";
-import { Negociacao } from "../models/negociacao.js";
-import { Negociacoes } from "../models/negociacoes.js";
-import { MensagemView } from "../views/mensagem-view.js";
-import { NegociacoesView } from "../views/negociacoes-view.js";
+import { DiasDaSemana } from '../enums/dias-da-semana.js';
+import { Negociacao } from '../models/negociacao.js';
+import { Negociacoes } from '../models/negociacoes.js';
+import { MensagemView } from '../views/mensagem-view.js';
+import { NegociacoesView } from '../views/negociacoes-view.js';
 
 export class NegociacaoController {
     private inputData: HTMLInputElement;
@@ -20,22 +20,28 @@ export class NegociacaoController {
     }
 
     public adiciona(): void {
+        /*
+            Zé, você já viu isso?
+        */
         const negociacao = Negociacao.criaDe(
-            this.inputData.value,
+            this.inputData.value, 
             this.inputQuantidade.value,
             this.inputValor.value
         );
+     
         if (!this.ehDiaUtil(negociacao.data)) {
-            this.mensagemView.update('Apenas negociações em dias úteis são aceitas');
+            this.mensagemView
+                .update('Apenas negociações em dias úteis são aceitas');
             return ;
         }
+
         this.negociacoes.adiciona(negociacao);
         this.limparFormulario();
         this.atualizaView();
     }
 
     private ehDiaUtil(data: Date) {
-        return data.getDay() > DiasDaSemana.DOMINGO
+        return data.getDay() > DiasDaSemana.DOMINGO 
             && data.getDay() < DiasDaSemana.SABADO;
     }
