+++ app/src/enums/dias-da-semana.ts
@@ -1,7 +1,7 @@
 export enum DiasDaSemana {
     DOMINGO = 0,
     SEGUNDA = 1,
-    TERÇA = 2,
+    TERCA = 2,
     QUARTA = 3,
     QUINTA = 4,
     SEXTA = 5,
