+++ app/src/models/negociacoes.ts
@@ -1,4 +1,4 @@
-import { Negociacao } from "./negociacao.js";
+import { Negociacao } from './negociacao.js';
 
 export class Negociacoes {
     private negociacoes: Negociacao[] = [];
