+++ app/src/views/mensagem-view.ts
@@ -1,10 +1,11 @@
-import { View } from "./view.js";
+
+import { View } from './view.js';
 
 export class MensagemView extends View<string> {
 
     protected template(model: string): string {
         return `
-        <p class="alert alert-info">${model}</p>
-        `;
+            <p class="alert alert-info">${model}</p>
+        `
     }
 }
