+++ app/src/views/view.ts
@@ -8,9 +8,8 @@ export abstract class View<T> {
         if (elemento) {
             this.elemento = elemento as HTMLElement;
         } else {
-            throw Error('Seletor não existe no DOM. Verifique.');
+            throw Error(`Seletor ${seletor} não existe no DOM. Verifique`);
         }
-
         if (escapar) {
             this.escapar = escapar;
         }
@@ -20,7 +19,7 @@ export abstract class View<T> {
         let template = this.template(model);
         if (this.escapar) {
             template = template
-                .replace(/<script>[\s\S]*?<script>/, '');
+                .replace(/<script>[\s\S]*?<\/script>/, '');
         }
         this.elemento.innerHTML = template;
     }
