+++ tsconfig.json
@@ -1,11 +1,11 @@
 {
     "compilerOptions": {
-        "outDir": "dist/js",
+        "outDir": "app/dist/js",
         "target": "ES6",
         "noEmitOnError": true,
         "noImplicitAny": true,
         "removeComments": true,
         "strictNullChecks": true
     },
-    "include": ["app/**/*"]
+    "include": ["app/src/**/*"]
 }
